worker primitives are SGX compatible, with all modules (including rpc)
