+++ enclave/Cargo.toml
@@ -57,8 +57,8 @@ webpki = { git = "https://github.com/mesalock-linux/webpki", branch = "mesalock_
 webpki-roots = { git = "https://github.com/mesalock-linux/webpki-roots", branch = "mesalock_sgx" }
 log = { git = "https://github.com/mesalock-linux/log-sgx" }
 env_logger = { git = "https://github.com/mesalock-linux/env_logger-sgx" }
-serde = { git = "https://github.com/mesalock-linux/serde-sgx" }
-serde_json = { git = "https://github.com/mesalock-linux/serde-json-sgx" }
+serde = { tag = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-sgx" }
+serde_json = { tag = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-json-sgx" }
 yasna = { rev = "sgx_1.1.3", default-features = false, features = ["bit-vec", "num-bigint", "chrono", "mesalock_sgx"], git = "https://github.com/mesalock-linux/yasna.rs-sgx" }
 rustls = { rev = "sgx_1.1.3", features = ["dangerous_configuration"], git = "https://github.com/mesalock-linux/rustls" }
 
@@ -86,8 +86,8 @@ test-utils = { path = "../primitives/test-utils", default-features = false, feat
 chain-relay = { path = "chain_relay" }
 substratee-stf = { path = "../stf", default-features = false, features = ["sgx"] }
 substratee-storage = { path = "../primitives/storage", default-features = false }
-substratee-worker-primitives= {path = "../primitives/worker", default-features = false }
-substratee-node-primitives = {path = "../primitives/node", default-features = false, features = ["sgx"] }
+substratee-worker-primitives= { path = "../primitives/worker", default-features = false }
+substratee-node-primitives = { path = "../primitives/node", default-features = false, features = ["sgx"] }
 
 # substrate deps
 frame-support = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
