+++ primitives/worker/Cargo.toml
@@ -7,10 +7,9 @@ edition = "2018"
 [dependencies]
 codec           = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive", "full"] }
 primitive-types = { version = "0.10.1", default-features = false, features = ["codec"] }
-serde           = { version = "1.0", optional = true}
-serde_derive    = { version = "1.0", optional = true}
-serde_json      = { version = "1.0", default-features = false, features = ["alloc"] }
 chrono          = { version = "0.4.19", default-features = false, features = ["alloc"]}
+serde           = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+serde_json      = { version = "1.0", default-features = false, features = ["alloc"] }
 
 # local deps
 substratee-storage = { path = "../storage", default-features = false }
@@ -21,18 +20,11 @@ sp-std = { version = "4.0.0-dev", default-features = false, git = "https://githu
 sp-core = { version = "4.0.0-dev", default-features = false, features = ["full_crypto"], git = "https://github.com/paritytech/substrate.git", branch = "master"}
 sp-runtime = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "master" }
 
-# sgx deps
-sgx_tstd = { rev = "v1.1.3", features = ["untrusted_fs","net","backtrace"], git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
-
-
-
-
 [features]
 default = ["std"]
 std = [ 'codec/std',
         'chrono/std',
-        'serde',
-        'serde_derive',
+        'serde/std',
         'serde_json/std',
         'primitive-types/std',
         'substratee-sidechain-primitives/std',
