+++ primitives/worker/src/rpc.rs
@@ -1,6 +1,27 @@
+extern crate alloc;
+
+use crate::DirectRequestStatus;
+use alloc::{borrow::ToOwned, string::String, vec::Vec};
 use codec::{Decode, Encode};
 use serde::{Deserialize, Serialize};
-use std::string::String;
+
+#[derive(Encode, Decode)]
+pub struct RpcReturnValue {
+	pub value: Vec<u8>,
+	pub do_watch: bool,
+	pub status: DirectRequestStatus,
+	//pub signature: Signature,
+}
+impl RpcReturnValue {
+	pub fn new(val: Vec<u8>, watch: bool, status: DirectRequestStatus) -> Self {
+		Self {
+			value: val,
+			do_watch: watch,
+			status,
+			//signature: sign,
+		}
+	}
+}
 
 #[derive(Clone, Encode, Decode, Serialize, Deserialize)]
 // Todo: result should not be Vec<u8>, but `T: Serialize`
